add forks count to viz preview
