+++ packages/neoFrontend/src/ForksLink/index.js
@@ -1,0 +1,1 @@
+export { ForksLink } from './ForksLink';
