+++ packages/neoFrontend/src/ForksLink/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { Link } from 'react-router-dom';
+
+export const VizLink = styled(Link)`
+  font-style: italic;
+`;
